chore: rename wrangler.nimbus-preview.json to wrangler.preview.json, build to dist
