+++ .github/workflows/publish-production.yml
@@ -57,7 +57,7 @@ jobs:
         continue-on-error: true
         with:
           name: site-html
-          path: dist-nimbus
+          path: dist
 
       - name: Deploy to Cloudflare Workers
         run: pnpm exec wrangler deploy
