+++ src/nimbus/tsconfig.json
@@ -13,5 +13,5 @@
 		"jsxImportSource": "react"
 	},
 	"include": ["../../.astro/types.d.ts", "**/*"],
-	"exclude": ["dist", "dist-nimbus"]
+	"exclude": ["dist"]
 }
