+++ tsconfig.json
@@ -10,5 +10,5 @@
 		"jsxImportSource": "react"
 	},
 	"include": [".astro/types.d.ts", "**/*"],
-	"exclude": ["dist", "dist-nimbus", "worker", ".flue", "src/nimbus"]
+	"exclude": ["dist", "worker", ".flue", "src/nimbus"]
 }
