--- wrangler.nimbus-preview.json
@@ -1,21 +1,0 @@
-{
-	"$schema": "./node_modules/wrangler/config-schema.json",
-	"name": "cloudflare-docs",
-	"account_id": "b54f07a6c269ecca2fa60f1ae4920c99",
-	"compatibility_date": "2025-06-02",
-	"compatibility_flags": ["nodejs_compat"],
-	"main": "./worker/index.preview.ts",
-	"rules": [
-		{ "type": "Text", "globs": ["**/__redirects"], "fallthrough": true }
-	],
-	"assets": {
-		"directory": "./dist-nimbus",
-		"binding": "ASSETS",
-		"not_found_handling": "404-page",
-		"run_worker_first": true
-	},
-	"r2_buckets": [{ "binding": "MIDDLECACHE", "bucket_name": "middlecache" }],
-	"observability": {
-		"enabled": true
-	}
-}
